+++ Configs/SilkyUIClientConfig.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using SilkyUIFramework.Helper;
 using Terraria.ModLoader.Config;
 
 namespace SilkyUIFramework.Configs;
