+++ Elements/UIElementGroup.cs
@@ -1,6 +1,4 @@
-﻿using SilkyUIFramework.Helper;
-
-namespace SilkyUIFramework.Elements;
+﻿namespace SilkyUIFramework.Elements;
 
 /// <summary>
 /// 似乎在密谋着什么，再等等...
